requirements.txt
requests
python-dotenv
pandas
mistralai